move ddl test
